fix 添加已下载书籍详情页，跳转到播放器页面

+++ module-listen/src/main/java/com/rm/module_listen/viewmodel/ListenSubsUpdateViewModel.kt
@@ -350,7 +350,7 @@ class ListenSubsUpdateViewModel : BaseVMViewModel() {
             path_url = model.path,
             play_count = model.play_count.toInt(),
             sequence = model.sequence.toInt(),
-            size = model.size.toInt(),
+            size = model.size.toLong(),
             created_at = "2020-09-09",
             recentPlay = 0L,
             chapter_id = model.chapter_id
